+++ fabric21/src/main/java/com/vecoo/extralib/gson/UtilGson.java
@@ -23,7 +23,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.function.Consumer;
 
-public abstract class UtilGson {
+public final class UtilGson {
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().disableHtmlEscaping().create();
 
     /**