+++ fabric21/src/main/java/com/vecoo/extralib/item/UtilItem.java
@@ -13,7 +13,7 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.component.CustomModelData;
 import org.jetbrains.annotations.NotNull;
 
-public class UtilItem {
+public final class UtilItem {
     /**
      * Parses an item ID into an {@link ItemStack}. If the item does not exist,
      * returns {@link ItemStack#EMPTY}.