+++ neoforge21/src/main/java/com/vecoo/extralib/chat/UtilChat.java
@@ -4,7 +4,7 @@ import com.vecoo.extralib.ExtraLib;
 import net.minecraft.network.chat.*;
 import org.jetbrains.annotations.NotNull;
 
-public class UtilChat {
+public final class UtilChat {
     /**
      * Formats a string into a Minecraft {@link Component} and replaces
      * color codes from '&' to '§'.
@@ -38,7 +38,7 @@ public class UtilChat {
      * @return a new {@link MutableComponent} with click action applied
      */
     @NotNull
-    public static MutableComponent clickableMessageCommand(@NotNull MutableComponent message, @NotNull String command) {
+    public static Component clickableMessageCommand(@NotNull Component message, @NotNull String command) {
         return message.copy().setStyle(Style.EMPTY.withClickEvent(
                 new ClickEvent(ClickEvent.Action.RUN_COMMAND, command)));
     }
@@ -64,7 +64,7 @@ public class UtilChat {
      * @return a new {@link MutableComponent} with click action applied
      */
     @NotNull
-    public static MutableComponent clickableMessageURL(@NotNull MutableComponent message, @NotNull String url) {
+    public static Component clickableMessageURL(@NotNull Component message, @NotNull String url) {
         return message.copy().setStyle(Style.EMPTY.withClickEvent(
                 new ClickEvent(ClickEvent.Action.OPEN_URL, url)));
     }
@@ -90,7 +90,7 @@ public class UtilChat {
      * @return a new {@link MutableComponent} with hover event applied
      */
     @NotNull
-    public static MutableComponent hoverMessageText(@NotNull MutableComponent message, @NotNull String text) {
+    public static Component hoverMessageText(@NotNull Component message, @NotNull String text) {
         return message.copy().setStyle(Style.EMPTY.withHoverEvent(
                 new HoverEvent(HoverEvent.Action.SHOW_TEXT, formatMessage(text))));
     }