+++ changelog.md
@@ -1,1 +1,4 @@
-* Change name methods.
+* Finalized classes.
+* Added shadow Spongepowered Configurate YAML.
+* Added class YamlConfigFactory.
+* Change ServerPlayer to Player, MutableComponent to Component.